Project submitted for eval
Updated the project with the files my teammate and I submitted for the evaluation of this school project.

This version of the project still has errors in the uv mapping of the mesh in the scenes "tcage" and "twindows". Also, the depth of focus is not implemented in this version.